requirements.txt
# Core dependencies
opencv-python>=4.8.0,<5.0.0
pyserial>=3.5,<4.0.0
pillow>=8.0.0,<11.0.0
numpy>=2.0.0,<2.3.0

# Video Processing Utilities
imutils>=0.5.4,<1.0.0

# Development tools (optional)
pytest>=6.0.0,<8.0.0
black>=21.0.0,<25.0.0